PART_Bonus/PPO/requirements.txt
"""
requirements.txt - All dependencies for Pokemon Red RL

Save this as 'requirements.txt' and install with:
    pip install -r requirements.txt

For GPU support, install PyTorch separately first:
    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
"""

# Core RL dependencies
torch>=2.0.0
gymnasium>=0.28.1
stable-baselines3>=2.0.0
tensorboard>=2.13.0

# Emulation
pyboy>=1.6.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Image processing
Pillow>=10.0.0
scikit-image>=0.21.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
psutil>=5.9.0
tqdm>=4.65.0

# Optional: Advanced monitoring
# wandb>=0.15.0  # Uncomment for Weights & Biases integration